Rename CustomEditorFactory -> CustomLayerTreeEditorFactory

+++ source/darefl/model/experimentaldatamodel.cpp
@@ -121,7 +121,7 @@ void ExperimentalDataModel::addDataToGroup(CanvasItem* data_group, RealDataStruc
     std::vector<double> axis_vec;
     std::vector<double> data_vec;
 
-    for (int i = 0; i < data_struct.data.size(); ++i) {
+    for (size_t i = 0; i < data_struct.data.size(); ++i) {
         if (!std::isnan(data_struct.axis.at(i)) && !std::isnan(data_struct.data.at(i))) {
             axis_vec.push_back(data_struct.axis.at(i));
             data_vec.push_back(data_struct.data.at(i));
